[identity] Refactor plugin definitions
